requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
cffi==1.14.5
dj-database-url>=0.5.0
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt>=4.4.0
django-storages==1.7.1
gunicorn==20.0.4
Pillow==8.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
whitenoise==4.1.4
